+++ src/views/index.js
@@ -42,6 +42,7 @@ export { default as tabs } from './tabs.js';
 export { default as tab } from './tab.js';
 export { default as text } from './text.js';
 export { default as textMatch } from './text-match.js';
+export { default as textNumeric } from './text-numeric.js';
 export { default as tocSection } from './toc-section.js';
 export { default as toggle } from './toggle.js';
 export { default as toggleGroup } from './toggle-group.js';