+++ src/views/text-numeric.usage.js
@@ -1,0 +1,12 @@
+export default {
+    demo: {
+        view: 'text-numeric',
+        data: '"Like a `text` view but adds a thousands separator to integer part of numbers, e.g. 12345678 or 12345.67890"'
+    },
+    examples: [
+        {
+            title: 'Shorthand usage',
+            view: 'text-numeric:1234567'
+        }
+    ]
+};